refactor : 테스트 리펙토링
