+++ src/test/kotlin/com/kotlin/boardproject/service/ConstSaver.kt
@@ -1,0 +1,4 @@
+package com.kotlin.boardproject.service
+
+const val nonExistPostId = -1L
+const val nonExistUserEmail = "fail@test.com"
